Move current interpolation from compile-time parameter to a command-line one
Add a new command-line option to set current interpolation.
Due to #3551, internal processing was already runtime.
Move current interpolation-related code to modern namespace and directory structure.

Current interpolation is no longer a template parameter of field solvers.
Adjust all .param files and cmakeFlags accordingly.
