+++ docs/TBG_macros.cfg
@@ -104,6 +104,11 @@ TBG_windowMovePoint="--windowMovePoint 0.9"
 # stop the moving window after given simulation step
 TBG_stopWindow="--stopWindow 1337"
 
+
+# Set current smoothing.
+# Supported values: none (default), binomial
+TBG_currentInterpolation="--currentInterpolation binomial"
+
 ################################################################################
 ## Placeholder for multi data plugins:
 ##
