+++ include/picongpu/fields/MaxwellSolver/ArbitraryOrderFDTD/ArbitraryOrderFDTD.def
@@ -19,7 +19,6 @@
 
 #pragma once
 
-#include "picongpu/fields/currentInterpolation/CurrentInterpolation.def"
 #include "picongpu/fields/differentiation/Curl.def"
 #include "picongpu/fields/MaxwellSolver/ArbitraryOrderFDTD/Derivative.def"
 #include "picongpu/fields/MaxwellSolver/Yee/Yee.def"
@@ -69,9 +68,9 @@ namespace picongpu
              *                     Same number of neighbors is used along all directions.
              *                     Order of derivative approximation is 2 * T_neighbors.
              */
-            template<uint32_t T_neighbors, typename T_CurrentInterpolation = currentInterpolation::None>
-            using ArbitraryOrderFDTD = ::picongpu::fields::maxwellSolver::
-                Yee<T_CurrentInterpolation, aoFDTD::CurlE<T_neighbors>, aoFDTD::CurlB<T_neighbors>>;
+            template<uint32_t T_neighbors>
+            using ArbitraryOrderFDTD
+                = ::picongpu::fields::maxwellSolver::Yee<aoFDTD::CurlE<T_neighbors>, aoFDTD::CurlB<T_neighbors>>;
 
             /* We need no definition of margins, because the Yee solver uses its curl
              * classes to define margins
