+++ include/picongpu/fields/MaxwellSolver/ArbitraryOrderFDTDPML/ArbitraryOrderFDTDPML.hpp
@@ -30,9 +30,8 @@ namespace pmacc
 {
     namespace traits
     {
-        template<uint32_t T_neighbors, typename T_CurrentInterpolation>
-        struct StringProperties<
-            ::picongpu::fields::maxwellSolver::ArbitraryOrderFDTDPML<T_neighbors, T_CurrentInterpolation>>
+        template<uint32_t T_neighbors>
+        struct StringProperties<::picongpu::fields::maxwellSolver::ArbitraryOrderFDTDPML<T_neighbors>>
         {
             static StringProperty get()
             {
