+++ include/picongpu/fields/MaxwellSolver/Lehe/Lehe.def
@@ -20,7 +20,6 @@
 
 #pragma once
 
-#include "picongpu/fields/currentInterpolation/CurrentInterpolation.def"
 #include "picongpu/fields/differentiation/Curl.def"
 #include "picongpu/fields/MaxwellSolver/Lehe/Derivative.def"
 #include "picongpu/fields/MaxwellSolver/Yee/Yee.def"
@@ -67,11 +66,8 @@ namespace picongpu
              * @tparam T_CherenkovFreeDir the direction (axis) which should be free of cherenkov radiation
              *                            0 = x, 1 = y, 2 = z
              */
-            template<
-                typename T_CurrentInterpolation = currentInterpolation::None,
-                uint32_t T_cherenkovFreeDir = lehe::CherenkovFreeDirection_Y>
-            using Lehe
-                = ::picongpu::fields::maxwellSolver::Yee<T_CurrentInterpolation, lehe::CurlE<T_cherenkovFreeDir>>;
+            template<uint32_t T_cherenkovFreeDir = lehe::CherenkovFreeDirection_Y>
+            using Lehe = ::picongpu::fields::maxwellSolver::Yee<lehe::CurlE<T_cherenkovFreeDir>>;
 
             /* We need no definition of margins, because the Yee solver uses its curl
              * classes to define margins
