+++ include/picongpu/fields/MaxwellSolver/None/None.def
@@ -20,7 +20,6 @@
 #pragma once
 
 #include <pmacc/math/Vector.hpp>
-#include "picongpu/fields/currentInterpolation/CurrentInterpolation.def"
 
 
 namespace picongpu
@@ -29,7 +28,6 @@ namespace picongpu
     {
         namespace maxwellSolver
         {
-            template<typename T_CurrentInterpolation = currentInterpolation::None>
             class None;
 
         } // namespace maxwellSolver
@@ -37,15 +35,15 @@ namespace picongpu
 
     namespace traits
     {
-        template<typename T_CurrentInterpolation>
-        struct GetMargin<picongpu::fields::maxwellSolver::None<T_CurrentInterpolation>, FIELD_B>
+        template<>
+        struct GetMargin<picongpu::fields::maxwellSolver::None, FIELD_B>
         {
             using LowerMargin = typename pmacc::math::CT::make_Int<simDim, 0>::type;
             using UpperMargin = LowerMargin;
         };
 
-        template<typename T_CurrentInterpolation>
-        struct GetMargin<picongpu::fields::maxwellSolver::None<T_CurrentInterpolation>, FIELD_E>
+        template<>
+        struct GetMargin<picongpu::fields::maxwellSolver::None, FIELD_E>
         {
             using LowerMargin = typename pmacc::math::CT::make_Int<simDim, 0>::type;
             using UpperMargin = LowerMargin;
