+++ include/picongpu/fields/MaxwellSolver/Yee/Yee.def
@@ -19,7 +19,6 @@
 
 #pragma once
 
-#include "picongpu/fields/currentInterpolation/CurrentInterpolation.def"
 #include "picongpu/fields/differentiation/Curl.def"
 #include "picongpu/fields/differentiation/Derivative.def"
 
@@ -37,10 +36,7 @@ namespace picongpu
 
             } // namespace yee
 
-            template<
-                typename T_CurrentInterpolation = currentInterpolation::None,
-                typename CurlE = yee::CurlRight,
-                typename CurlB = yee::CurlLeft>
+            template<typename CurlE = yee::CurlRight, typename CurlB = yee::CurlLeft>
             class Yee;
 
         } // namespace maxwellSolver
