+++ include/picongpu/fields/MaxwellSolver/Yee/Yee.hpp
@@ -42,12 +42,11 @@ namespace picongpu
     {
         namespace maxwellSolver
         {
-            template<typename T_CurrentInterpolation, class CurlE, class CurlB>
+            template<class CurlE, class CurlB>
             class Yee
             {
             private:
                 typedef MappingDesc::SuperCellSize SuperCellSize;
-
 
                 std::shared_ptr<FieldE> fieldE;
                 std::shared_ptr<FieldB> fieldB;
@@ -64,7 +63,7 @@ namespace picongpu
                     PMACC_CASSERT_MSG(
                         Courant_Friedrichs_Levy_condition_failure____check_your_grid_param_file,
                         (SPEED_OF_LIGHT * SPEED_OF_LIGHT * DELTA_T * DELTA_T * INV_CELL2_SUM) <= 1.0
-                            && sizeof(T_CurrentInterpolation*) != 0);
+                            && sizeof(SuperCellSize*) != 0);
 
                     typedef SuperCellDescription<
                         SuperCellSize,
@@ -103,7 +102,6 @@ namespace picongpu
 
             public:
                 using CellType = cellType::Yee;
-                using CurrentInterpolation = T_CurrentInterpolation;
 
                 Yee(MappingDesc cellDescription) : m_cellDescription(cellDescription)
                 {
@@ -154,15 +152,15 @@ namespace picongpu
 
     namespace traits
     {
-        template<typename T_CurrentInterpolation, class CurlE, class CurlB>
-        struct GetMargin<picongpu::fields::maxwellSolver::Yee<T_CurrentInterpolation, CurlE, CurlB>, FIELD_B>
+        template<class CurlE, class CurlB>
+        struct GetMargin<picongpu::fields::maxwellSolver::Yee<CurlE, CurlB>, FIELD_B>
         {
             using LowerMargin = typename CurlB::LowerMargin;
             using UpperMargin = typename CurlB::UpperMargin;
         };
 
-        template<typename T_CurrentInterpolation, class CurlE, class CurlB>
-        struct GetMargin<picongpu::fields::maxwellSolver::Yee<T_CurrentInterpolation, CurlE, CurlB>, FIELD_E>
+        template<class CurlE, class CurlB>
+        struct GetMargin<picongpu::fields::maxwellSolver::Yee<CurlE, CurlB>, FIELD_E>
         {
             using LowerMargin = typename CurlE::LowerMargin;
             using UpperMargin = typename CurlE::UpperMargin;
