+++ include/picongpu/fields/currentInterpolation/CurrentInterpolation.hpp
@@ -19,8 +19,8 @@
 
 #pragma once
 
-#include "picongpu/fields/currentInterpolation/None/None.hpp"
-#include "picongpu/fields/currentInterpolation/Binomial/Binomial.hpp"
+#include "picongpu/fields/currentInterpolation/Binomial.hpp"
+#include "picongpu/fields/currentInterpolation/None.hpp"
 
 #include <pmacc/math/Vector.hpp>
 #include <pmacc/traits/GetStringProperties.hpp>
@@ -28,66 +28,64 @@
 
 namespace picongpu
 {
-    namespace currentInterpolation
+    namespace fields
     {
-        /** Singleton to represent current interpolation kind
-         *
-         * It does not perform interpolation itself, that is done by functors None and Binomial.
-         * Provides run-time utilities to get margin values and string properties.
-         *
-         * Note: for now it is called CurrentInterpolationInfo to not conflict with CurrentInterpolation type alias
-         * used in standard .param files. Will be renamed to just CurrentInterpolation after a transition to a
-         * run-time parameter
-         */
-        struct CurrentInterpolationInfo
+        namespace currentInterpolation
         {
-        public:
-            //! Supported interpolation kinds
-            enum class Kind
+            /** Singleton to represent current interpolation kind
+             *
+             * It does not perform interpolation itself, that is done by functors None and Binomial.
+             * Provides run-time utilities to get margin values and string properties.
+             */
+            struct CurrentInterpolation
             {
-                None,
-                Binomial
+            public:
+                //! Supported interpolation kinds
+                enum class Kind
+                {
+                    None,
+                    Binomial
+                };
+
+                //! Interpolation kind used in the simulation
+                Kind kind = Kind::Binomial;
+
+                //! Get the single instance of the current interpolation object
+                static CurrentInterpolation& get()
+                {
+                    static CurrentInterpolation instance;
+                    return instance;
+                }
+
+                //! Get string properties
+                static pmacc::traits::StringProperty getStringProperties()
+                {
+                    return get().kind == Kind::None ? None::getStringProperties() : Binomial::getStringProperties();
+                }
+
+                //! Get the lower margin of the used interpolation functor
+                static pmacc::math::Vector<int, simDim> getLowerMargin()
+                {
+                    return get().kind == Kind::None ? None::LowerMargin::toRT() : Binomial::LowerMargin::toRT();
+                }
+
+                //! Get the upper margin of the used interpolation functor
+                static pmacc::math::Vector<int, simDim> getUpperMargin()
+                {
+                    return get().kind == Kind::None ? None::UpperMargin::toRT() : Binomial::UpperMargin::toRT();
+                }
+
+                //! Copy construction is forbidden
+                CurrentInterpolation(CurrentInterpolation const&) = delete;
+
+                //! Assignment is forbidden
+                CurrentInterpolation& operator=(CurrentInterpolation const&) = delete;
+
+            private:
+                CurrentInterpolation() = default;
+                ~CurrentInterpolation() = default;
             };
 
-            //! Interpolation kind used in the simulation
-            Kind kind = Kind::None;
-
-            //! Get the single instance of the current interpolation object
-            static CurrentInterpolationInfo& get()
-            {
-                static CurrentInterpolationInfo instance;
-                return instance;
-            }
-
-            //! Get string properties
-            static pmacc::traits::StringProperty getStringProperties()
-            {
-                return get().kind == Kind::None ? None::getStringProperties() : Binomial::getStringProperties();
-            }
-
-            //! Get the lower margin of the used interpolation functor
-            static pmacc::math::Vector<int, simDim> getLowerMargin()
-            {
-                return get().kind == Kind::None ? None::LowerMargin::toRT() : Binomial::LowerMargin::toRT();
-            }
-
-            //! Get the upper margin of the used interpolation functor
-            static pmacc::math::Vector<int, simDim> getUpperMargin()
-            {
-                return get().kind == Kind::None ? None::UpperMargin::toRT() : Binomial::UpperMargin::toRT();
-            }
-
-            //! Copy construction is forbidden
-            CurrentInterpolationInfo(CurrentInterpolationInfo const&) = delete;
-
-            //! Assignment is forbidden
-            CurrentInterpolationInfo& operator=(CurrentInterpolationInfo const&) = delete;
-
-        private:
-            CurrentInterpolationInfo() = default;
-            ~CurrentInterpolationInfo() = default;
-        };
-
-    } // namespace currentInterpolation
-
+        } // namespace currentInterpolation
+    } // namespace fields
 } // namespace picongpu
