+++ include/picongpu/param/fieldSolver.param
@@ -23,9 +23,6 @@
  *
  * Select the numerical Maxwell solver (e.g. Yee's method).
  *
- * Also allows to configure ad hoc mitigations for high frequency
- * noise in some setups via current smoothing.
- *
  * \attention
  * Currently, the laser initialization in PIConGPU is implemented to work with the standard Yee solver.
  * Using a solver of higher order will result in a slightly increased laser amplitude and energy than expected.
@@ -35,48 +32,31 @@
 #pragma once
 
 #include "picongpu/fields/MaxwellSolver/Solvers.def"
-#include "picongpu/fields/currentInterpolation/CurrentInterpolation.def"
 
 
 namespace picongpu
 {
     namespace fields
     {
-        /** Current Interpolation
-         *
-         * CurrentInterpolation is used to set a method performing the
-         * interpolate/assign operation from the generated currents of particle
-         * species to the electro-magnetic fields.
-         *
-         * Allowed values are:
-         *   - None:
-         *     - default for staggered grids/Yee-scheme
-         *     - updates E
-         *   - Binomial: 2nd order Binomial filter
-         *     - smooths the current before assignment in staggered grid
-         *     - updates E & breaks local charge conservation slightly
-         */
-        using CurrentInterpolation = currentInterpolation::None;
-
         /** FieldSolver
          *
-         * Field Solver Selection:
-         *  - Yee< CurrentInterpolation > : Standard Yee solver approximating derivatives with respect to time and
+         * Field Solver Selection (note <> for some solvers):
+         *  - Yee<> : Standard Yee solver approximating derivatives with respect to time and
          * space by second order finite differences.
-         *  - YeePML< CurrentInterpolation >: Standard Yee solver using Perfectly Matched Layer Absorbing Boundary
+         *  - YeePML<>: Standard Yee solver using Perfectly Matched Layer Absorbing Boundary
          * Conditions (PML)
-         *  - Lehe< CurrentInterpolation >: Num. Cherenkov free field solver in a chosen direction
-         *  - LehePML< CurrentInterpolation >: Num. Cherenkov free field solver in a chosen direction
-         *                                     using Perfectly Matched Layer Absorbing Boundary Conditions (PML)
-         *  - ArbitraryOrderFDTD< 4, CurrentInterpolation >: Solver using 4 neighbors to each direction to approximate
+         *  - Lehe<>: Num. Cherenkov free field solver in a chosen direction
+         *  - LehePML<>: Num. Cherenkov free field solver in a chosen direction
+         *               using Perfectly Matched Layer Absorbing Boundary Conditions (PML)
+         *  - ArbitraryOrderFDTD<4>: Solver using 4 neighbors to each direction to approximate
          * *spatial* derivatives by finite differences. The number of neighbors can be changed from 4 to any positive,
          * integer number. The order of the solver will be twice the number of neighbors in each direction. Yee's
          * method is a special case of this using one neighbor to each direction.
-         *  - ArbitraryOrderFDTDPML< 4, CurrentInterpolation >: ArbitraryOrderFDTD solver using Perfectly Matched Layer
-         *                                                      Absorbing Boundary Conditions (PML)
-         *  - None< CurrentInterpolation >: disable the vacuum update of E and B
+         *  - ArbitraryOrderFDTDPML<4>: ArbitraryOrderFDTD solver using Perfectly Matched Layer
+         *                              Absorbing Boundary Conditions (PML)
+         *  - None: disable the vacuum update of E and B
          */
-        using Solver = maxwellSolver::Yee<CurrentInterpolation>;
+        using Solver = maxwellSolver::Yee<>;
 
     } // namespace fields
 } // namespace picongpu
