+++ include/picongpu/plugins/adios/WriteMeta.hpp
@@ -291,7 +291,7 @@ namespace picongpu
 
                 writeMeta::OfAllSpecies<>()(threadParams, fullMeshesPath);
 
-                GetStringProperties<currentInterpolation::CurrentInterpolationInfo> currentSmoothingProp;
+                GetStringProperties<fields::currentInterpolation::CurrentInterpolation> currentSmoothingProp;
                 const std::string currentSmoothing(currentSmoothingProp["name"].value);
                 ADIOS_CMD(adios_define_attribute_byvalue(
                     threadParams->adiosGroupHandle,
