+++ include/picongpu/simulation/stage/CurrentInterpolationAndAdditionToEMF.hpp
@@ -33,10 +33,11 @@
 #include <pmacc/type/Area.hpp>
 
 #include <boost/mpl/count.hpp>
+#include <boost/program_options/options_description.hpp>
 
 #include <cstdint>
 #include <stdexcept>
-#include <type_traits>
+#include <string>
 
 
 namespace picongpu
@@ -51,6 +52,18 @@ namespace picongpu
             class CurrentInterpolationAndAdditionToEMF
             {
             public:
+                /** Register program options for current interpolation
+                 *
+                 * @param desc program options following boost::program_options::options_description
+                 */
+                void registerHelp(po::options_description& desc)
+                {
+                    desc.add_options()(
+                        "currentInterpolation",
+                        po::value<std::string>(&kindName),
+                        (std::string("Current interpolation kind [None, Binomial] default: " + kindName).c_str()));
+                }
+
                 /** Initialize the current interpolation stage
                  *
                  * This method has to be called during initialization of the simulation.
@@ -58,15 +71,15 @@ namespace picongpu
                  */
                 void init()
                 {
-                    // Convert compile-time current interpolation set in the field solver to a runtime value
-                    using namespace currentInterpolation;
-                    auto& interpolation = CurrentInterpolationInfo::get();
-                    if(std::is_same<fields::Solver::CurrentInterpolation, None>::value)
-                        interpolation.kind = CurrentInterpolationInfo::Kind::None;
-                    else if(std::is_same<fields::Solver::CurrentInterpolation, Binomial>::value)
-                        interpolation.kind = CurrentInterpolationInfo::Kind::Binomial;
+                    using namespace fields::currentInterpolation;
+                    auto& interpolation = CurrentInterpolation::get();
+                    // So far there are only two kinds and so names are hardcoded
+                    if(kindName == "none")
+                        interpolation.kind = CurrentInterpolation::Kind::None;
+                    else if(kindName == "binomial")
+                        interpolation.kind = CurrentInterpolation::Kind::Binomial;
                     else
-                        throw std::runtime_error("Unsupported current interpolation type used in the field solver");
+                        throw std::runtime_error("Unsupported current interpolation type");
                 }
 
                 /** Compute the current created by particles and add it to the current density
@@ -86,7 +99,7 @@ namespace picongpu
                     DataConnector& dc = Environment<>::get().DataConnector();
                     auto& fieldJ = *dc.get<FieldJ>(FieldJ::getName(), true);
                     auto eRecvCurrent = fieldJ.asyncCommunication(__getTransactionEvent());
-                    auto& interpolation = currentInterpolation::CurrentInterpolationInfo::get();
+                    auto& interpolation = fields::currentInterpolation::CurrentInterpolation::get();
                     auto const currentRecvLower = interpolation.getLowerMargin();
                     auto const currentRecvUpper = interpolation.getUpperMargin();
 
@@ -116,6 +129,9 @@ namespace picongpu
                 }
 
             private:
+                //! Name set by program option
+                std::string kindName = "none";
+
                 /* Call addCurrentToEMF method of fieldJ for the given area
                  *
                  * This function performs a transition from the run-time realm of CurrentInterpolation into the
@@ -128,12 +144,12 @@ namespace picongpu
                 template<std::uint32_t T_area>
                 void addCurrentToEMF(FieldJ& fieldJ) const
                 {
-                    using currentInterpolation::CurrentInterpolationInfo;
-                    auto const kind = CurrentInterpolationInfo::get().kind;
-                    if(kind == CurrentInterpolationInfo::Kind::None)
-                        fieldJ.addCurrentToEMF<T_area>(currentInterpolation::None{});
+                    using namespace fields::currentInterpolation;
+                    auto const kind = CurrentInterpolation::get().kind;
+                    if(kind == CurrentInterpolation::Kind::None)
+                        fieldJ.addCurrentToEMF<T_area>(None{});
                     else
-                        fieldJ.addCurrentToEMF<T_area>(currentInterpolation::Binomial{});
+                        fieldJ.addCurrentToEMF<T_area>(Binomial{});
                 }
             };
