+++ share/picongpu/tests/compileCurrentSolver/cmakeFlags
@@ -33,7 +33,7 @@ flags[0]="-DPARAM_OVERWRITES:LIST='-DPARAM_CURRENTSOLVER=EmZ<UsedParticleShape>;
 flags[1]="-DPARAM_OVERWRITES:LIST='-DPARAM_CURRENTSOLVER=EmZ<UsedParticleShape>;-DPARAM_PARTICLESHAPE=PCS;-DPARAM_DIMENSION=DIM2'"
 # Esirkepov and TSC is tested in most examples
 flags[2]="-DPARAM_OVERWRITES:LIST='-DPARAM_CURRENTSOLVER=Esirkepov<UsedParticleShape>;-DPARAM_PARTICLESHAPE=CIC'"
-flags[3]="-DPARAM_OVERWRITES:LIST='-DPARAM_CURRENTSOLVER=Esirkepov<UsedParticleShape>;-DPARAM_PARTICLESHAPE=TSC;-DPARAM_CURRENTINTERPOLATION=Binomial'"
+flags[3]="-DPARAM_OVERWRITES:LIST='-DPARAM_CURRENTSOLVER=Esirkepov<UsedParticleShape>;-DPARAM_PARTICLESHAPE=TSC'"
 flags[4]="-DPARAM_OVERWRITES:LIST='-DPARAM_CURRENTSOLVER=Esirkepov<UsedParticleShape>;-DPARAM_PARTICLESHAPE=PQS;-DPARAM_DIMENSION=DIM2'"
 flags[5]="-DPARAM_OVERWRITES:LIST='-DPARAM_CURRENTSOLVER=VillaBune<>;-DPARAM_PARTICLESHAPE=CIC'"
